+++ src/main/scala/coulomb/Quantity.scala
@@ -33,82 +33,6 @@ import test._
 
 
 
-trait CanonicalSig[U] {
-  type Out
-  def coef: Rational
-}
-
-object CanonicalSig {
-  type Aux[U, O] = CanonicalSig[U] { type Out = O }
-
-  implicit def witnessUnitlessCM: Aux[Unitless, HNil] = {
-    new CanonicalSig[Unitless] {
-      type Out = HNil
-      val coef = Rational(1)
-    }
-  }
-
-  implicit def witnessBaseUnitCM[U](implicit buU: BaseUnit[U]): Aux[U, (U, XInt1) :: HNil] = {
-    new CanonicalSig[U] {
-      type Out = (U, XInt1) :: HNil
-      val coef = Rational(1)
-    }
-  }
-
-  implicit def witnessDerivedUnitCM[U, D, DC](implicit du: DerivedUnit[U, D], dm: Aux[D, DC]): Aux[U, DC] = {
-    new CanonicalSig[U] {
-      type Out = DC
-      val coef = du.coef * dm.coef
-    }
-  }
-
-  implicit def witnessMulCM[L, LC, R, RC, OC](implicit l: Aux[L, LC], r: Aux[R, RC], u: UnifySigMul.Aux[LC, RC, OC]): Aux[%*[L, R], OC] = {
-    new CanonicalSig[%*[L, R]] {
-      type Out = OC
-      val coef = l.coef * r.coef
-    }
-  }
-
-  implicit def witnessDivCM[L, LC, R, RC, OC](implicit l: Aux[L, LC], r: Aux[R, RC], u: UnifySigDiv.Aux[RC, LC, OC]): Aux[%/[L, R], OC] = {
-    new CanonicalSig[%/[L, R]] {
-      type Out = OC
-      val coef = l.coef / r.coef
-    }
-  }
-
-  implicit def witnessPowCM[B, BC, E, OC](implicit b: Aux[B, BC], u: ApplySigPow.Aux[E, BC, OC], e: singleton.ops.Id[E]): Aux[%^[B, E], OC] = {
-    new CanonicalSig[%^[B, E]] {
-      type Out = OC
-      val coef = b.coef.pow(e.value.asInstanceOf[Int])
-    }
-  }
-}
-
-trait StandardSig[U] {
-  type Out
-}
-object StandardSig {
-  type Aux[U, O] = StandardSig[U] { type Out = O }
-
-  implicit def witnessUnitlessCM: Aux[Unitless, HNil] =
-    new StandardSig[Unitless] { type Out = HNil }
-
-  implicit def witnessBaseUnitCM[U](implicit buU: BaseUnit[U]): Aux[U, (U, XInt1) :: HNil] =
-    new StandardSig[U] { type Out = (U, XInt1) :: HNil }
-
-  implicit def witnessDerivedUnitCM[U](implicit du: DerivedUnit[U, _]): Aux[U, (U, XInt1) :: HNil] =
-    new StandardSig[U] { type Out = (U, XInt1) :: HNil }
-
-  implicit def witnessMulCM[L, LC, R, RC, OC](implicit l: Aux[L, LC], r: Aux[R, RC], u: UnifySigMul.Aux[LC, RC, OC]): Aux[%*[L, R], OC] =
-    new StandardSig[%*[L, R]] { type Out = OC }
-
-  implicit def witnessDivCM[L, LC, R, RC, OC](implicit l: Aux[L, LC], r: Aux[R, RC], u: UnifySigDiv.Aux[RC, LC, OC]): Aux[%/[L, R], OC] =
-    new StandardSig[%/[L, R]] { type Out = OC }
-
-  implicit def witnessPowCM[B, BC, E, OC](implicit b: Aux[B, BC], u: ApplySigPow.Aux[E, BC, OC], e: singleton.ops.Id[E]): Aux[%^[B, E], OC] =
-    new StandardSig[%^[B, E]] { type Out = OC }
-}
-
 // assuming inputs (U, P) are already from some canonical sig; so a type U will never pre-exist in M
 trait InsertSortedUnitSig[U, P, M] {
   type Out
